+++ node/src/types/kernel/api/filetree/getHPathByID/payload.d.ts
@@ -15,14 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-/* directories */
-export * as createDocWithMd from "./createDocWithMd";
-export * as getHPathByID from "./getHPathByID";
-export * as getHPathByPath from "./getHPathByPath";
-export * as removeDoc from "./removeDoc";
-export * as renameDoc from "./renameDoc";
-
-/* files */
-export * as listDocsByPath from "./listDocsByPath";
-export * as searchDocs from "./searchDocs";
+/**
+ * Get human readable path based on ID
+ */
+export interface IPayload {
+    /**
+     * document ID
+     */
+    id: string;
+}